Update Index.cshtml
The pagination parameters were adjusted. Added CSS class superuser checkbox to "IsSuperUser" checkbox (chkIsSuperuser). Introduced JavaScript code to disable the "IsSuperUser" checkbox based on the user's superuser status (isCurrentUserSuperUser). And the values of the filters were changed to be able to use them with the new method implemented using the UserController class.