Update Upendo.UserManager.dnn
Added definition and configuration files that are involved in the module installation process. The '01.00.01.SqlDataProvider' file has been added to the module structure, to ensure the correct installation of the 'UUM_GetUsers' stored procedure in the database during the module installation process. And Uninstall.SqlDataProvider for the module uninstallation process.